+++ examples/form/src/ElementForm.vue
@@ -50,6 +50,7 @@ const { config, data } = createFormData({
 
 const rules = getElementPlusRules(new Pension())
 const submitForm = (formEl: any) => {
+  console.log(data.value)
   if (!formEl)
     return
   formEl.$.exposed?.validate((valid: boolean) => {